# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Tutor API",
    "version": "1.0.0",
    "description": "AI-powered learning platform backend API.\n\n**Modules:** Auth, Organizations, Academic Structure, Materials, Lessons, Flashcards, Quizzes, Chat (RAG), Progress, Analytics, Search, Notifications, Usage, Jobs\n\n**Response format:** `{ success, message, data }`"
  },
  "servers": [{ "url": "http://localhost:7001", "description": "Development" }],
...
  "paths": {
...
    "/materials/upload/pdf": {
      "post": {
        "tags": ["Materials"],
        "summary": "Upload a PDF material",
        "description": "Uploads PDF to Cloudflare R2 and queues async AI processing (extract → chunk → embed → Qdrant index).",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file", "organizationId", "topicId", "title"],
                "properties": {
                  "file": { "type": "string", "format": "binary", "description": "PDF file (max 50MB)" },
                  "organizationId": { "type": "string" },
                  "topicId": { "type": "string", "description": "Topic this material belongs to" },
                  "title": { "type": "string", "example": "Biology Chapter 3" },
                  "description": { "type": "string" },
                  "subjectName": { "type": "string", "example": "Biology" },
                  "tags": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "PDF uploaded — processing queued",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Material" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "503": { "description": "R2 storage or Redis queue unavailable" }
        }
      }
    },
    "/materials/upload/text": {
      "post": {
        "tags": ["Materials"],
        "summary": "Upload text content as material",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["organizationId", "topicId", "title", "content"],
                "properties": {
                  "organizationId": { "type": "string" },
                  "topicId": { "type": "string" },
                  "title": { "type": "string", "example": "Photosynthesis Notes" },
                  "content": { "type": "string", "example": "Photosynthesis is the process..." },
                  "description": { "type": "string" },
                  "subjectName": { "type": "string" },
                  "tags": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Text uploaded — processing queued" }
        }
      }
    },
    "/materials/youtube": {
      "post": {
        "tags": ["Materials"],
        "summary": "Add a YouTube video as material",
        "description": "Fetches video transcript in background and queues AI processing.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["organizationId", "topicId", "title", "youtubeUrl"],
                "properties": {
                  "organizationId": { "type": "string" },
                  "topicId": { "type": "string" },
                  "title": { "type": "string", "example": "Khan Academy - Algebra" },
                  "youtubeUrl": {
                    "type": "string",
                    "example": "https://www.youtube.com/watch?v=VIDEO_ID"
                  },
                  "description": { "type": "string" },
                  "subjectName": { "type": "string" },
                  "tags": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "YouTube link added — processing queued" }
        }
      }
    },
    "/materials": {
      "get": {
        "tags": ["Materials"],
        "summary": "List materials for an organization",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "organizationId", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "topicId", "in": "query", "schema": { "type": "string" } },
          { "name": "subjectId", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          {
            "name": "processingStatus",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["PENDING", "QUEUED", "PROCESSING", "COMPLETED", "FAILED"]
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": { "type": "string", "enum": ["PDF", "TEXT", "YOUTUBE"] }
          }
        ],
        "responses": {
          "200": { "description": "Materials retrieved" }
        }
      }
    },
    "/materials/{id}": {
      "get": {
        "tags": ["Materials"],
        "summary": "Get material by ID",
        "description": "Returns metadata including processingStatus, summary, and chunkCount.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Material retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Material" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Materials"],
        "summary": "Delete material and associated vectors",
        "description": "Soft-deletes material, removes R2 file and Qdrant vectors.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Material deleted" }
        }
      }
    },
    "/materials/{id}/chunks": {
      "get": {
        "tags": ["Materials"],
        "summary": "Get processed text chunks",
        "description": "Available after processingStatus is COMPLETED.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Chunks retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/MaterialChunk" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": { "description": "Processing not complete" }
        }
      }
    },
    "/materials/{id}/reprocess": {
      "post": {
        "tags": ["Materials"],
        "summary": "Re-queue material for AI processing",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Reprocessing queued" }
        }
      }
    },
    "/materials/{id}/archive": {
      "patch": {
        "tags": ["Materials"],
        "summary": "Archive material",
        "description": "Sets material status to INACTIVE without deleting vectors.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Material archived" }
        }
      }
    },
    "/materials/{id}/processing-logs": {
      "get": {
        "tags": ["Materials"],
        "summary": "Get material processing logs",
        "description": "Returns stage-by-stage processing history (UPLOADED → CHUNKING → EMBEDDING → etc.).",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Processing logs retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/MaterialProcessingLog" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/teacher/materials": { "get": { "tags": ["Teacher Portal"], "summary": "Teacher materials with processing status", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "organizationId", "in": "query", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "OK" } } } },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT access token from /auth/login"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string", "example": "Success" },
          "data": { "type": "object" }
        }
      },
...
      "Material": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "organizationId": { "type": "string" },
          "topicId": { "type": "string" },
          "subjectId": { "type": "string" },
          "academicYearId": { "type": "string" },
          "uploadedBy": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "type": { "type": "string", "enum": ["PDF", "TEXT", "YOUTUBE"] },
          "processingStatus": {
            "type": "string",
            "enum": ["PENDING", "QUEUED", "PROCESSING", "COMPLETED", "FAILED"]
          },
          "processingStage": {
            "type": "string",
            "enum": ["UPLOADED", "PROCESSING", "CHUNKING", "EMBEDDING", "GENERATING", "COMPLETED", "FAILED"],
            "description": "Detailed pipeline stage for processing tracking"
          },
          "fileName": { "type": "string" },
          "mimeType": { "type": "string" },
          "fileSize": { "type": "number" },
          "r2Url": { "type": "string" },
          "sourceUrl": { "type": "string" },
          "summary": { "type": "string" },
          "chunkCount": { "type": "number" },
          "errorMessage": { "type": "string" },
          "jobId": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "subjectName": { "type": "string" },
          "processingStartedAt": { "type": "string", "format": "date-time" },
          "processingCompletedAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "MaterialChunk": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "chunkIndex": { "type": "number" },
          "content": { "type": "string" },
          "startChar": { "type": "number" },
          "endChar": { "type": "number" }
        }
      },
...
      "MaterialProcessingLog": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "materialId": { "type": "string" },
          "organizationId": { "type": "string" },
          "stage": {
            "type": "string",
            "enum": ["UPLOADED", "PROCESSING", "CHUNKING", "EMBEDDING", "GENERATING", "COMPLETED", "FAILED"]
          },
          "message": { "type": "string" },
          "metadata": { "type": "object" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}